started making changes to verilator bb stuff. tests are failing... not sure why
